refactor: update _CLI import
