+++ src/taipy/core/_version/_cli/_version_cli.py
@@ -11,7 +11,7 @@
 
 import sys
 
-from taipy.config._cli._cli import _CLI
+from taipy.cli import _CLI
 from taipy.logger._taipy_logger import _TaipyLogger
 
 from ...exceptions.exceptions import VersionIsNotProductionVersion
